[libfoil] Allow to choose the dictionary for Base64 input
FOIL_INPUT_BASE64_STANDARD and FOIL_INPUT_BASE64_FILESAFE flags can
be used to force the use of a specific dictionary table (referred to
by RFC 4648 as "base64" and "base64url", respectively). If none or
both of those flags are specified, the table is picked automatically
based on the input, if only one flag is given, only that table is used.